tests/wpt/web-platform-tests/2dcontext/path-objects/2d.path.rect.winding.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: 2d.path.rect.winding</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="stylesheet" href="/common/canvas-tests.css">
<body class="show_output">

<h1>2d.path.rect.winding</h1>
<p class="desc"></p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

ctx.fillStyle = '#0f0';
ctx.fillRect(0, 0, 100, 50);
ctx.beginPath();
ctx.fillStyle = '#f00';
ctx.rect(0, 0, 50, 50);
ctx.rect(100, 50, -50, -50);
ctx.rect(0, 25, 100, -25);
ctx.rect(100, 25, -100, 25);
ctx.fill();
_assertPixel(canvas, 25,12, 0,255,0,255, "25,12", "0,255,0,255");
_assertPixel(canvas, 75,12, 0,255,0,255, "75,12", "0,255,0,255");
_assertPixel(canvas, 25,37, 0,255,0,255, "25,37", "0,255,0,255");
_assertPixel(canvas, 75,37, 0,255,0,255, "75,37", "0,255,0,255");


});
</script>
